final/requirements.txt
colorgram.py==1.2.0
deap==1.3.1
matplotlib==3.2.1
opencv_python==4.2.0.34
numpy==1.18.4
Katna==0.4.1.4
Pillow==7.2.0
